1.完成V0.1.4, 实现读取数据库单表读取; 2.bugfix, 修正long字段错误问题

+++ eppdev-jee-cg/src/main/java/cn/eppdev/jee/cg/utils/GeneratorUtils.java
@@ -151,7 +151,7 @@ public class GeneratorUtils {
     public static List<ColumnTypeInfo> listAllColumnType(String dbType){
         List<ColumnTypeInfo> list = new ArrayList<>();
         list.add(new ColumnTypeInfo("int", "Integer", "integer"));
-        list.add(new ColumnTypeInfo("long", "Long", "long"));
+        list.add(new ColumnTypeInfo("long", "Integer", "long"));
         list.add(new ColumnTypeInfo("decimal", "java.math.BigDecimal", "double"));
         list.add(new ColumnTypeInfo("char", "String", "char"));
         list.add(new ColumnTypeInfo("varchar", "String", "varchar"));
